Update tour.md (#3323)
## **Description**
Renaming from "tour of jaseci" to "key features"

I added a part 1 that is describing the key features conceptually
without detailed examples. This is for readers to quickly get the key
value-adds without getting distracted by detailed code snippets, which
can be quite long.


part 2 will be code snippets. which need to be updated in a separate PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>